requirements.txt
streamlit
python-dotenv
tavily-python
langchain
langchain-core